refactor: add visually-hidden span and change imports tm-71
