+++ apps/frontend/src/libs/components/icon/icon.tsx
@@ -1,5 +1,6 @@
+import { type IconName } from "~/libs/types/types.js";
+
 import { iconNameToSvg } from "./libs/maps/maps.js";
-import { type IconName } from "./libs/types/types.js";
 
 type Properties = {
 	name: IconName;
