+++ apps/frontend/src/libs/types/icon-name.type.ts
