+++ apps/frontend/src/libs/types/types.ts
@@ -1,4 +1,5 @@
 export { type AsyncThunkConfig } from "./async-thunk-config.type.js";
+export { type IconName } from "./icon-name.type.js";
 export {
 	type CourseDto,
 	type ServerErrorDetail,
